Clean up repository and Add flowers to the code
Since the support for legacy garden flowers is
terrible I imported the code from matplotlib
backend and navigationdrawer flowers into my own
code. Also a lot of modification that I never
commited back in March are in this commit.
Added requirements to make it easier to install
them in the future and made a virtual env to
separate concerns.